frontend: Moved TestInfoType to TestConfig.ts
