+++ frontend/src/features/tests/BDITest/layout/BDITestContent.tsx
@@ -1,6 +1,6 @@
-import TestInfoType from "@/features/tests/TestInfoType";
 import BDITest from "@/features/tests/BDITest/types/BDITest";
 import BDIQuestionCard from "@/features/tests/BDITest/components/BDIQuestionCard";
+import {TestInfoType} from "@/features/tests/TestConfig";
 
 export default function BDITestContent({ test, disabled }: TestInfoType<BDITest>) {
     return (
