+++ frontend/src/features/tests/BDITest/layout/BDITestHeader.tsx
@@ -1,9 +1,9 @@
-import TestInfoType from "@/features/tests/TestInfoType";
 import BDITest from "@/features/tests/BDITest/types/BDITest";
 import getBDITestMetrics from "@/features/tests/BDITest/utils/getBDITestMetrics";
 import { useMemo } from "react";
 import {Typography} from "@mui/material";
 import {Roles} from "@/schemas/Role";
+import {TestInfoType} from "@/features/tests/TestConfig";
 
 export default function BDITestHeader({test, role}: TestInfoType<BDITest>) {
     const {
