+++ frontend/src/features/tests/PCL5Test/layout/PCL5Content.tsx
@@ -1,12 +1,12 @@
 "use client";
 
-import TestInfoType from "@/features/tests/TestInfoType";
 import {PCL5Test} from "@/features/tests/PCL5Test/types/PCL5Test";
 import getTestMetrics from "@/features/tests/PCL5Test/utils/getTestMetrics";
 import {useMemo} from "react";
 import {Typography} from "@mui/material";
 import {Roles} from "@/schemas/Role";
 import PCL5QuestionCard from "@/features/tests/PCL5Test/components/PCL5QuestionCard";
+import {TestInfoType} from "@/features/tests/TestConfig";
 
 export function PCL5ContentHeader({test, role}: TestInfoType<PCL5Test>) {
     const {
